fix :: saga 단계 줄이기(anime 서버에 대한 접근 제거)
