+++ src/main/java/windeath44/orchestration/adapter/in/messaging/MemorialApplicationApproveSubscribe.java
@@ -1,6 +1,7 @@
 package windeath44.orchestration.adapter.in.messaging;
 
 import com.example.avro.CharacterAvroSchema;
+import com.example.avro.MemorialAvroSchema;
 import lombok.RequiredArgsConstructor;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Component;
@@ -11,8 +12,8 @@ import windeath44.orchestration.domain.port.in.MemorialApplicationApproveUseCase
 public class MemorialApplicationApproveSubscribe {
     private final MemorialApplicationApproveUseCase memorialApplicationApproveUseCase;
 
-    @KafkaListener(topics="character-memorialized-response", groupId = "memorial")
-    public void listen(CharacterAvroSchema message) {
+    @KafkaListener(topics="memorial-creation-response", groupId = "memorial")
+    public void listen(MemorialAvroSchema message) {
       memorialApplicationApproveUseCase.execute(message);
     }
 }
