--- src/main/java/windeath44/orchestration/application/service/AnimeService.java
@@ -1,16 +1,0 @@
-package windeath44.orchestration.application.service;
-
-import com.example.avro.MemorialAvroSchema;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import windeath44.orchestration.domain.port.out.CharacterEventPublisher;
-
-@Service
-@RequiredArgsConstructor
-public class AnimeService {
-  private final CharacterEventPublisher characterMemorializingEventPublisher;
-
-  public void memorial(MemorialAvroSchema memorialAvroSchema) {
-    characterMemorializingEventPublisher.publish(memorialAvroSchema);
-  }
-}
