+++ src/main/java/windeath44/orchestration/domain/mapper/EventMapper.java
@@ -64,31 +64,4 @@ public class EventMapper {
   }
 
 
-  public CharacterEvent characterEvent(CharacterAvroSchema characterAvroSchema) {
-    String aggregateId = "character-" + characterAvroSchema.getCharacterId();
-    String aggregateType = "ANIME";
-    EventType eventType = EventType.CHARACTER_MEMORIALIZED;
-
-    return CharacterEvent.builder()
-            .aggregateId(aggregateId)
-            .aggregateType(aggregateType)
-            .eventType(eventType)
-            .eventData(characterAvroSchema)
-            .build();
-  }
-
-  public CharacterEvent characterCompensateEvent(CharacterAvroSchema characterAvroSchema) {
-    String aggregateId = "character-" + characterAvroSchema.getCharacterId();
-    String aggregateType = "ANIME";
-
-    EventType eventType = EventType.CHARACTER_MEMORIALIZED_FAILED;
-
-    return CharacterEvent.builder()
-            .aggregateId(aggregateId)
-            .aggregateType(aggregateType)
-            .eventType(eventType)
-            .eventData(characterAvroSchema)
-            .build();
-  }
-
 }
