--- src/main/java/windeath44/orchestration/domain/port/in/CharacterMemorializingUseCase.java
@@ -1,8 +1,0 @@
-package windeath44.orchestration.domain.port.in;
-
-
-import com.example.avro.MemorialAvroSchema;
-
-public interface CharacterMemorializingUseCase {
-  void execute(MemorialAvroSchema memorialAvroSchema);
-}
