+++ src/main/java/windeath44/orchestration/domain/port/in/MemorialApplicationApproveUseCase.java
@@ -1,7 +1,7 @@
 package windeath44.orchestration.domain.port.in;
 
-import com.example.avro.CharacterAvroSchema;
+import com.example.avro.MemorialAvroSchema;
 
 public interface MemorialApplicationApproveUseCase {
-  void execute(CharacterAvroSchema characterAvroSchema);
+  void execute(MemorialAvroSchema memorialAvroSchema);
 }
