--- src/main/java/windeath44/orchestration/domain/port/in/MemorialDeleteUseCase.java
@@ -1,7 +1,0 @@
-package windeath44.orchestration.domain.port.in;
-
-import com.example.avro.CharacterAvroSchema;
-
-public interface MemorialDeleteUseCase {
-  void execute(CharacterAvroSchema characterAvroSchema);
-}
